Splitting primitive object test suites
This change splits the original set of primitive object test suites into
separate modules. Minor refactoring to account for structure changes is
included.
